Operations type has been change to double (instead deciimal)

+++ Core/Calculator.Application.Models/Operation/Calculate/OperationCalculateDto.cs
@@ -2,8 +2,8 @@
 {
     public class OperationCalculateDto
     {
-        public decimal LeftOperand { get; set; }
+        public double LeftOperand { get; set; }
 
-        public decimal RightOperand { get; set; }
+        public double RightOperand { get; set; }
     }
 }
